add clean code
