+++ src/router/user.ts
@@ -4,7 +4,5 @@ const userRouter: Router = Router();
 import * as userController from "../controller/user";
 
 userRouter.get("/", userController.getAllUsers);
-userRouter.post("/login", userController.getLogin);
-userRouter.post("/register", userController.getRegister);
 
 export default userRouter;
